templates/etl-python/etl-storage-cf/{{cookiecutter.name}}/requirements.txt
google-cloud-storage==2.7.0
python-dotenv==1.0.1
